Refactor nav controls to separate file
